rename mapping section -> mappint template
