+++ static/js/import_wizard2/components/mapping_templates/multi_var_mulit_date_col.tsx
@@ -16,8 +16,8 @@
 
 import React from "react";
 
-import { MappingSectionProps } from "../../templates";
+import { MappingTemplateProps } from "../../templates";
 
-export function SingleVarMultiDateCol(props: MappingSectionProps): JSX.Element {
+export function MultiVarMultiDateCol(props: MappingTemplateProps): JSX.Element {
   return <></>;
 }
